+++ DemoSnippets/ToolboxInteractionLogic.cs
@@ -23,6 +23,7 @@ namespace DemoSnippets
         private const string DefaultTabName = "Demo";
         private const string ToolboxDataSourceId = "DemoSnippets";
         private const string ToolboxDataLabel = "DSLabel";
+        private const string ToolboxDataFileName = "DSFileName";
 
         private readonly AsyncPackage package;
 
@@ -58,7 +59,7 @@ namespace DemoSnippets
                     item.Tab = DefaultTabName;
                 }
 
-                await AddToToolboxAsync(item.Tab, item.Label, item.Snippet);
+                await AddToToolboxAsync(item.Tab, item.Label, item.Snippet, Path.GetFileNameWithoutExtension(fileName));
                 addedCount += 1;
             }
 
@@ -286,7 +287,7 @@ namespace DemoSnippets
             return result;
         }
 
-        private static async Task<IDataObject> AddToToolboxAsync(string tab, string label, string actualText)
+        private static async Task<IDataObject> AddToToolboxAsync(string tab, string label, string actualText, string sourceFileName)
         {
             await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync(CancellationToken.None);
 
@@ -308,9 +309,18 @@ namespace DemoSnippets
                 // Set label here so easy to read back and use in logging when removing.
                 tbItem.SetData(ToolboxDataLabel, label);
 
-                await OutputPane.Instance.WriteAsync($"Adding '{label}' to tab '{tab}'");
+                // Store this so can avoid duplicating toolbox entries when repeatedly loading the same file
+                tbItem.SetData(ToolboxDataFileName, sourceFileName);
 
-                // TODO: avoid adding duplicate items
+                if (await RemoveExistingItemAsync(tab, label, sourceFileName))
+                {
+                    await OutputPane.Instance.WriteAsync($"Replacing '{label}' in tab '{tab}'");
+                }
+                else
+                {
+                    await OutputPane.Instance.WriteAsync($"Adding '{label}' to tab '{tab}'");
+                }
+
                 toolbox?.AddItem(tbItem, itemInfo, tab);
 
                 return tbItem;
@@ -320,6 +330,68 @@ namespace DemoSnippets
                 await OutputPane.Instance.WriteAsync($"Error: {e.Message}{Environment.NewLine}{e.Source}{Environment.NewLine}{e.StackTrace}");
                 return null;
             }
+        }
+
+        private static async Task<bool> RemoveExistingItemAsync(string tabName, string label, string originalSourceFile)
+        {
+            await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync(CancellationToken.None);
+
+            try
+            {
+                var toolbox = await Instance.ServiceProvider.GetServiceAsync(typeof(IVsToolbox)) as IVsToolbox;
+
+                IEnumToolboxTabs tbTabs = null;
+                toolbox?.EnumTabs(out tbTabs);
+                var returnedTabNames = new string[1];
+
+                while (tbTabs?.Next(1, returnedTabNames, out uint tabsReturned) == VSConstants.S_OK)
+                {
+                    if (tabName == returnedTabNames[0])
+                    {
+                        IEnumToolboxItems tbItems = null;
+
+                        toolbox?.EnumItems(tabName, out tbItems);
+
+                        var dataObjects = new IDataObject[1];
+                        uint fetched = 0;
+
+                        while (tbItems?.Next(1, dataObjects, out fetched) == VSConstants.S_OK)
+                        {
+                            if (dataObjects[0] != null && fetched == 1)
+                            {
+                                var itemDataObject = new OleDataObject(dataObjects[0]);
+
+                                if (itemDataObject.ContainsText(TextDataFormat.Text))
+                                {
+                                    var isDemoSnippet = itemDataObject.GetData(ToolboxDataSourceId);
+
+                                    if (isDemoSnippet != null)
+                                    {
+                                        var itemFile = itemDataObject.GetData(ToolboxDataFileName).ToString();
+
+                                        if (itemFile == originalSourceFile)
+                                        {
+                                            var itemLabel = itemDataObject.GetData(ToolboxDataLabel).ToString();
+
+                                            if (itemLabel == label)
+                                            {
+                                                toolbox?.RemoveItem(dataObjects[0]);
+                                                return true;
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            catch (Exception e)
+            {
+                await OutputPane.Instance.WriteAsync($"Error: {e.Message}{Environment.NewLine}{e.Source}{Environment.NewLine}{e.StackTrace}");
+            }
+
+            return false;
         }
     }
 }